split up pdb tools
