+++ mdagent/tools/base_tools/__init__.py
@@ -10,7 +10,7 @@ from .preprocess_tools.clean_tools import (
     SpecializedCleanTool,
 )
 from .preprocess_tools.packing import PackMolTool
-from .preprocess_tools.pdb_tools import ProteinName2PDBTool, SmallMolPDB, get_pdb
+from .preprocess_tools.pdb_get import ProteinName2PDBTool, SmallMolPDB, get_pdb
 from .simulation_tools.create_simulation import ModifyBaseSimulationScriptTool
 from .simulation_tools.setup_and_run import (
     InstructionSummary,
